Oresteia: Use example namespace on local fact node IDs
Using CASE's namespace in the `@context` `@vocab` made all non-prefixed
terms be considered as CASE ontology terms.  Unfortunately, this also
captured the knowledge base terms (e.g.
`investigation-4586742a-710a-454f-bcb8-b60e230ec1b2`) in the sweeping
declaration of things that belong to the CASE ontology namespace.  With
a pending mass-influx of UCO namespaces for this example file, this
patch begins the practice of explicitly identifying the namespace of
all ontology vocabulary *and* knowledge base identifiers.

The default namespace for objects in the example knowledge base needs to
be distinct from the ontology namespace(s).  I suggest `urn:example:`
(RFC 6963) as the base of sample, non-network-resolving prefixes.
https://tools.ietf.org/html/rfc6963

Note that syntax requirements of URN specify that `#` characters cannot
be used.
https://tools.ietf.org/html/rfc8141#section-2.2

This patch implements `url:example:kb:` as the blank prefix for
knowledge base facts in the Oresteia example.  To assist with
programmatic checks for unintentionally unprefixed identifier strings,
the "local" namespace as the context `@vocab` quietly annotates missed
strings.

Note this patch does not correct the namespacing of CASE and UCO
ontology terms.  That will be a heftier diff.