Make all IDs into URIs
When the examples were written, the `@id` keys in the JSON dictionaries
were accidentally written as literal strings.  Per the RDF spec, the
subject of a triple must be an IRI or blank [1].  As originally drafted,
the ids as literal strings meant triples could not be defined.  The
curious-looking symptom is that triples with these incorrect subjects
would not load into an RDF engine, so translating to, say, Turtle would
cause only the blank nodes of the property bundles to be translated,
attached to nothing.

These `@id` keys were fixed with the following commands, run on the
whitespace-normalized JSON:

    sed --in-place --expression 's/"@id": "/"@id": ":/' *.json
    sed --in-place --expression 's/"::/":/' *.json

(The second command undoes any second-applications of this fix.)

[1] https://www.w3.org/TR/rdf11-concepts/#section-triples