Update some prototype ontology references to UCO 0.3.0
This patch focuses revisions on example files that included storage
system ontology terms, mainly things related to files and file
characteristics.  This is *part* of the work needed to migrate the
examples to UCO 0.3.0.  This specific part was done to assist with
testing a CASE implementation that focuses on translating storage system
metadata.

A summary of the changes:
* Migration of some prototype terms to UCO 0.3.0.
* Some corrections of the conversion from CASE 0.1's `Relationship` to
  UCO 0.3.0's `CyberRelationship`.
* Refinements of literal-type designations, such as
  `uco-observable:sizeInBytes` being denoted as `xsd:long`.
* Assignments of literal-type designations for some relationship
  enumerants.
* Removal of some unused namespace prefixes (e.g. `olo`).

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>