Add development guidelines to README
This patch enacts the committee decisions in CASE Jira Story ONT-140.

Please note this disclaimer related to mention of specific software:
Participation by NIST in the creation of the instructions to this
toolkit is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software or toolkit is necessarily the best
available for the purpose.