Select ChangeLog format
As part of the UCO and CASE Shared Development Practices, a top-level
file `/ChangeLog` is to be tracked in the ontology repository.  This
file serves a purpose between the Git history and a release's Release
Notes, by providing a concise summary of Git merges made into the
develop branch, and by providing a running list of changes to document
in the Release Notes.

This addresses CASE Story ONT-288, Change Proposal 11.

Following the period of fast-track review of the accompanying Change
Proposal, a follow-on commit will link the Change Proposal document and
feedback.

References:
* [ONT-288] (CP-11) CASE needs a format for the ontology repository's
  ChangeLog