Add infrastructure to run normalization as unit test
This patch adds Makefiles that confirm CASE ontology Turtle files are
normalized according to the command specified in NORMALIZING.md.

Usage:
* `make` - does nothing.  Nothing in the ontology currently depends on
  executing code to build a component.
* `make check` - downloads `rdf-toolkit.jar` to the `/lib` directory in
  this repository, if not already present, and checks that ontology
  files match the output of `rdf-toolkit.jar` running against them.
  - `make check`, or `make .lib.done.log`, needs to be called in the
    root directory of this repository at least once.  After that,
    `make check` will work in any directory at or under `/ontology`.

The hash of `rdf-toolkit.jar` will be recorded in the Git history in the
next commit, with a timestamp indicating when I first encountered this
hash.

This patch begins a resolution of ONT-251.

References:
* [ONT-251] Establish unit tests, possibly Continuous Integration, for
  ontology repository

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>