Base CASE-specific imports on versionIRI property
This patch does not do the same for UCO imports, because UCO needs to
review a parallel proposal to CASE's CP-26 in order to confirm its
serving domain.  Hence, this patch is a partial implementation of
ONT-412 - necessary changes, but not yet sufficient changes.

rdf-toolkit.jar was used to normalize the files just before committing,
which induced one added import-comment line in investigation-da.ttl.

References:
* [ONT-412] (CP-26) Implement versionIRI in CASE 0.2.0
* [UCO OC-101] (CP-40) Implement versionIRI in UCO 0.4.0

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>