Run typing script
The vocabulary.ttl file had one change added that was rejected, changing
the xsd: prefix to xs:.  This is a known issue with rdflib providing
`xs:` as a default prefix when the XML Schema Datatypes are not used in
the ontology.  (This was previously addressed in AC-154.)

This script run used the virtual environment built in the tracked UCO
submodule.

References:
* [AC-154] XML Schema datatype prefix needs to be explicit in all
  examples
* [ONT-456] (CP-41) Convert current property restrictions and domain
  assertions to SHACL class shapes