Add testing build chain, with XFAIL and XPASS test data and validation output
Note that pytest will fail unless a pySHACL release >0.15.0 is used.
A follow-on patch will add a git submodule as scaffolding, to be
removed before release.

Also note that, in contrast to UCO's similarly structured tests,
a `*_PASS_validation.ttl` graph is recorded even when it records a
validation failure.  `pytest` is left to report the validation error.

References:
* [ONT-456] (CP-41) Convert current property restrictions and domain
  assertions to SHACL class shapes