Add testing build chain, with XFAIL and XPASS test data and validation output
Note that, in contrast to UCO's similarly structured tests,
a `*_PASS_validation.ttl` graph is recorded even when it records a
validation failure.  `pytest` is left to review and report the
validation error.

The script `test_validation.py` has type annotations added, and has
undergone static source review with `mypy`.  It will be left for future
consideration to add static source review to the ontology CI.

References:
* [ONT-456] (CP-41) Convert current property restrictions and domain
  assertions to SHACL class shapes

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>