Delegate rdf-toolkit provisioning to UCO build process
This patch removes all handling of rdf-toolkit.jar from the CASE
ontology repository.  UCO had adopted CASE's rdf-toolkit retrieval in
version 0.7.0, and now UCO Issue 373 will be upgrading the retrieval
process.

This patch handles delegation of rdf-toolkit provisioning.  A follow-on
patch will handle the rdf-toolkit and Java upgrade that is the focus of
UCO Issue 373.

References:
* https://github.com/ucoProject/UCO/issues/373

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>