Adapt for UCO files being under different directories
This is a preparation for UCO Issue 387, where a third level of
subdirectory depth was added.

Before the UCO Git submodule pointer is updated to incorporate
UCO Issue 387's resolution, the `/bin/ls` command will emit that one of
its shell wildcards didn't work, and the command will exit 0.  This does
not fail the Make process.

To avoid complexities with Git merging, the UCO submodule pointer is not
updated in this patch.

The Makefile shell recipe declines to use `find` due to potentially
incorporating CI artifacts.

References:
* https://github.com/ucoProject/UCO/issues/387

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>