Confirm an InvestigativeAction results in at least one ProvenanceRecord
This new shape stemmed from discussion on CASE Issue 136.

As a matter of preserving backwards compatibility, this patch introduces
the shape requiring `ProvenanceRecord`s with a `sh:Warning`-level
severity.  In CASE 2.0.0, this requirement will be strengthened into a
`sh:Violation`.

A separate proposal will be filed with UCO to test the minimum qualified
cardinality OWL structure.  A draft of that syntax review system was
used to test this patch.

This patch adds a version floor for pySHACL to ensure an update in
qualified value shape handling is included, which is necessary for the
new property shape to function when using pySHACL.

Disclaimer:

References:
* https://github.com/RDFLib/pySHACL/issues/213
* https://github.com/casework/CASE/issues/136
* https://github.com/casework/CASE/issues/146

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>