Revise Technician example to have person be the action-performer
Note this patch does not tie the role to the action, nor does it define
an investigation to tie the action, person, role, relationship, etc.,
together, due to needing to address broader design issues pertaining to
`uco-role:Role`, some of which are under exploration in CASE-Examples
Issue 155.

This patch also fixes a typo'd concept - `uco-core:Identity` was used
instead of `uco-identity:Identity`.
This was not revised to `uco-identity:Person` in light of discussion
going on in UCO Issue 652.

No effects were observed on Make-managed files.

References:
* https://github.com/casework/CASE/issues/184
* https://github.com/casework/CASE-Examples/issues/155
* https://github.com/ucoProject/UCO/issues/652

Signed-off-by: Alex Nelsin <alexander.nelson@nist.gov>